Started to work on functional test for the library.
Fixed problem where its possible to acknowledge a message in the channel that it wasn't sent from.
Now have it where channels are created per a consumer.
Consumer closes the channel when it is removed.
Changed the api where you create consumer from the connection instead of the channel for threading reason.